prettifyed and made service/client to one edge only
